create DbSets

+++ EntityFramework/Entities/Category.cs
@@ -11,7 +11,7 @@ namespace EntityFramework.Entities
         
             public int Id { get; set; }
             public string Name { get; set; }
-            public decimal Summ { get; set; }
+        public int Summ { get; set; }
             public virtual ICollection<Cost> Costs{ get; set; }
             public int AcountId { get; set; }
             public virtual Acount Acount { get; set; }
